moved to util file
