+++ src/main/kotlin/net/ccbluex/liquidbounce/features/command/commands/creative/CommandItemEnchant.kt
@@ -25,14 +25,11 @@ import net.ccbluex.liquidbounce.features.command.builder.ParameterBuilder
 import net.ccbluex.liquidbounce.utils.client.chat
 import net.ccbluex.liquidbounce.utils.client.mc
 import net.ccbluex.liquidbounce.utils.client.regular
+import net.ccbluex.liquidbounce.utils.item.addEnchantment
+import net.ccbluex.liquidbounce.utils.item.clearEnchantments
 import net.ccbluex.liquidbounce.utils.item.isNothing
-import net.minecraft.enchantment.Enchantment
-import net.minecraft.enchantment.EnchantmentHelper
+import net.ccbluex.liquidbounce.utils.item.removeEnchantment
 import net.minecraft.item.ItemStack
-import net.minecraft.item.Items
-import net.minecraft.nbt.NbtCompound
-import net.minecraft.nbt.NbtElement
-import net.minecraft.nbt.NbtList
 import net.minecraft.network.packet.c2s.play.CreativeInventoryActionC2SPacket
 import net.minecraft.registry.Registries
 import net.minecraft.util.Hand
@@ -243,30 +240,7 @@ object CommandItemEnchant {
         }
     }
 
-//    private fun getNbtlEnchantmentslist(item: ItemStack) =
 
 
 
-    private fun addEnchantment(item: ItemStack, enchantment: Enchantment, level: Int?) {
-        val nbt = item.orCreateNbt
-        if (nbt?.contains(ItemStack.ENCHANTMENTS_KEY, NbtElement.LIST_TYPE.toInt()) == false) {
-            nbt.put(ItemStack.ENCHANTMENTS_KEY, NbtList())
-        }
-        val nbtList = nbt?.getList(ItemStack.ENCHANTMENTS_KEY, NbtElement.COMPOUND_TYPE.toInt())
-        nbtList?.add(EnchantmentHelper.createNbt(EnchantmentHelper.getEnchantmentId(enchantment), level ?: enchantment.maxLevel))
-    }
-
-    private fun removeEnchantment(item: ItemStack, enchantment: Enchantment){
-        val nbt = item.nbt ?: return
-        if (!nbt.contains(ItemStack.ENCHANTMENTS_KEY, NbtElement.LIST_TYPE.toInt())) {
-            return
-        }
-        val nbtList = nbt.getList(ItemStack.ENCHANTMENTS_KEY, NbtElement.COMPOUND_TYPE.toInt())
-        nbtList.removeIf { (it as NbtCompound).getString("id") == EnchantmentHelper.getEnchantmentId(enchantment).toString() }
-    }
-
-    private fun clearEnchantments(item: ItemStack) {
-        val nbt = item.nbt ?: return
-        nbt.remove(ItemStack.ENCHANTMENTS_KEY)
-    }
 }
